Fix Serving of Media Files
Modify the `/media` location block in the nginx configuration so that
the block is always used when the URI prefix matches `/media`. This
fixes an issue where nginx would search the top-level `root` directory
for media files because they matched the static assets regex location
block further down in the configuration file.